# api-contract.yml
openapi: 3.0.0
info:
  title: CAVE API
  version: 1.0.0
  description: API para o sistema CAVE - Sistema de gerenciamento de produtos e compras
  contact:
    name: CAVE Team
    email: support@cave.com
servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento
  - url: https://api.cave.com
    description: Servidor de produção
paths:
  /api/login:
    post:
      tags:
        - Autenticação
      summary: Realizar login
      description: Autentica um usuário e retorna um token JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: Nome de usuário
                  example: "admin"
                password:
                  type: string
                  description: Senha do usuário
                  example: "admin123"
                  format: password
      responses:
        "200":
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token JWT para autenticação
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          description: Dados obrigatórios não fornecidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint /api/login
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: "Erro ao processar requisição"
...